+++ src/views/OutputView.js
@@ -1,5 +1,7 @@
+//@ts-check
+
 import { LOTTO_RANK, OUTPUT_MESSAGES } from '../lib/constants.js';
-import { calculateMatchCount } from '../lib/utils.js';
+import { calculateMatchCount } from '../util/calculator.js';
 
 class OutputView {
   static #print(message) {
